Crate cfx_tasks

Modules

shutdown

Helper for shutdown signals

Structs

PanickedTaskError

Error with the name of the task that panicked and an error downcasted to string, if possible.

TaskExecutor

A type that can spawn new tokio tasks

TaskManager

Many reth components require to spawn tasks for long-running jobs. For example discovery spawns tasks to handle egress and ingress of udp traffic or network that spawns session tasks that handle the traffic to and from a peer.

TokioTaskExecutor

An TaskSpawner that uses [tokio::task::spawn] to execute tasks

Traits

TaskSpawner

TaskSpawnerExt

TaskSpawner with extended behaviour