cfx_tasks

Struct TaskManager

pub struct TaskManager { /* private fields */ }

Many reth components require to spawn tasks for long-running jobs. For example discovery spawns tasks to handle egress and ingress of udp traffic or network that spawns session tasks that handle the traffic to and from a peer.

To unify how tasks are created, the TaskManager provides access to the configured Tokio runtime. A TaskManager stores the [tokio::runtime::Handle] it is associated with. In this way it is possible to configure on which runtime a task is executed.

The main purpose of this type is to be able to monitor if a critical task panicked, for diagnostic purposes, since tokio task essentially fail silently. Therefore, this type is a Stream that yields the name of panicked task, See TaskExecutor::spawn_critical. In order to execute Tasks use the TaskExecutor type TaskManager::executor.

Implementations

impl TaskManager

pub fn current() -> Self

Returns a TaskManager over the currently running Runtime.

Panics

This will panic if called outside the context of a Tokio runtime.

pub fn new(handle: Handle) -> Self

Create a new instance connected to the given handle’s tokio runtime.

pub fn executor(&self) -> TaskExecutor

Returns a new TaskExecutor that can spawn new tasks onto the tokio runtime this type is connected to.

pub fn graceful_shutdown(self)

Fires the shutdown signal and awaits until all tasks are shutdown.

pub fn graceful_shutdown_with_timeout(self, timeout: Duration) -> bool

Fires the shutdown signal and awaits until all tasks are shutdown.

Returns true if all tasks were shutdown before the timeout elapsed.